fixed handling of zip files without inner folder structure
The following scipt expected the files in a subfolder with the same name as the zip file. or it would stop executing. The path was created incorrectly if the zip does not contain a folder structure.

Furthermore, the zip could not be deleted since it was still opened, if it did not contain a folder structure.